requirements.txt
opencv-contrib-python
matplotlib
pillow
pillow-heif